+++ src/functions/emulator/stop-firebase-storage.ts
@@ -1,4 +1,3 @@
-import { WorkspaceContext } from '@causa/workspace';
 import { EmulatorStop } from '@causa/workspace-core';
 import { DockerEmulatorService } from '@causa/workspace-core/services';
 import {
@@ -10,14 +9,16 @@ import {
  * Implements {@link EmulatorStop} for the Firebase Storage emulator.
  */
 export class EmulatorStopForFirebaseStorage extends EmulatorStop {
-  async _call(context: WorkspaceContext): Promise<string> {
-    const containerName = getFirebaseStorageContainerName(context);
+  async _call(): Promise<string> {
+    const containerName = getFirebaseStorageContainerName(this._context);
 
-    context.logger.info('️🍱 Stopping Firebase Storage emulator.');
+    this._context.logger.info('️🍱 Stopping Firebase Storage emulator.');
 
-    await context.service(DockerEmulatorService).stop(containerName);
+    await this._context.service(DockerEmulatorService).stop(containerName);
 
-    context.logger.info('️🍱 Successfully stopped Firebase Storage emulator.');
+    this._context.logger.info(
+      '️🍱 Successfully stopped Firebase Storage emulator.',
+    );
 
     return FIREBASE_STORAGE_EMULATOR_NAME;
   }