+++ src/functions/event-topic/broker-publish-events-google.call.ts
@@ -1,4 +1,3 @@
-import type { WorkspaceContext } from '@causa/workspace';
 import type { PublishOptions } from '@google-cloud/pubsub';
 import { FlowControlledPublisher } from '@google-cloud/pubsub/build/src/publisher/flow-publisher.js';
 import { PubSubService } from '../../services/index.js';
@@ -26,14 +25,13 @@ const PUBLISH_OPTIONS: PublishOptions = {
 
 export default async function call(
   this: EventTopicBrokerPublishEventsForGoogle,
-  context: WorkspaceContext,
 ): Promise<void> {
-  const publisher: FlowControlledPublisher = context
+  const publisher: FlowControlledPublisher = this._context
     .service(PubSubService)
     .pubSub.topic(this.topicId, PUBLISH_OPTIONS)
     .flowControlled();
 
-  context.logger.info('📫 Publishing events.');
+  this._context.logger.info('📫 Publishing events.');
 
   let numEvents = 0;
   for await (const event of this.source()) {
@@ -42,21 +40,20 @@ export default async function call(
     const wait = publisher.publish({
       data: event.data,
       attributes: event.attributes,
-      orderingKey: event.key,
     });
     if (wait) {
-      context.logger.debug(
+      this._context.logger.debug(
         'Waiting for the publisher to catch up before resuming publishing.',
       );
       await wait;
     }
 
     if (numEvents % PUBLISH_PROGRESS_LOG_INTERVAL === 0) {
-      context.logger.debug(`Published ${numEvents} events.`);
+      this._context.logger.debug(`Published ${numEvents} events.`);
     }
   }
 
   await publisher.all();
 
-  context.logger.info(`📫 Finished publishing ${numEvents} events.`);
+  this._context.logger.info(`📫 Finished publishing ${numEvents} events.`);
 }