+++ src/functions/secret/fetch-access-token.call.ts
@@ -1,4 +1,3 @@
-import type { WorkspaceContext } from '@causa/workspace';
 import { GoogleApisService } from '../../services/google-apis.js';
 import {
   AuthClientResponseError,
@@ -7,9 +6,10 @@ import {
 
 export default async function call(
   this: SecretFetchForGoogleAccessToken,
-  context: WorkspaceContext,
 ): Promise<string> {
-  const authClient = await context.service(GoogleApisService).getAuthClient();
+  const authClient = await this._context
+    .service(GoogleApisService)
+    .getAuthClient();
   const { token } = await authClient.getAccessToken();
 
   if (!token) {